Update MysqliDb.php
Fixed lost precision with the conversion from float to int.

+++ src/MysqliDb.php
@@ -47,7 +47,7 @@ class MysqliDb {
   /* --------------- *
      Other Variables
    * --------------- */
-  protected int $traceStartQ;
+  protected float $traceStartQ;
   protected bool $traceEnabled = TRUE;
   protected null|string $traceStripPrefix = '';
